# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "VGCA NextCloud",
        "version": "1.0.0"
    },
    "paths": {
        "/api/key-encryption": {
            "get": {
                "tags": [
                    "keyEncryption"
                ],
                "description": "User uid",
                "responses": {
                    "200": {
                        "description": "Trả về  tất cả key encrytion"
                    }
                }
            },
            "put": {
                "tags": [
                    "keyEncryption"
                ],
                "description": "Cập nhật thông tin key encrytion",
                "operationId": "updateKeyEncrytion",
                "requestBody": {
                    "description": "Update user object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateKeyEncryptionSchema"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Trả về true hoặc false thể hiện thao tác thành công hay không"
                    }
                }
            },
            "post": {
                "tags": [
                    "keyEncryption"
                ],
                "description": "Lưu key encrytion đối với mỗi user và room",
                "operationId": "createKeyEncrytion",
                "requestBody": {
                    "description": "Create user object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateKeyEncryptionSchema"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Trả về  key encrytion",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/KeyEncryptionSchema"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/key-encryption/rooms/{room_id}": {
            "get": {
                "tags": [
                    "keyEncryption"
                ],
                "description": "room uid",
                "operationId": "2bb001a3513d74b16908c9c57aeb8bf6",
                "parameters": [
                    {
                        "name": "room_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "acca7421-d920-4751-b227-a21ef604f655"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Trả về  tất cả key encrytion của 1 room"
                    }
                }
            },
            "delete": {
                "tags": [
                    "keyEncryption"
                ],
                "description": "Xoá tất cả Key của 1 room",
                "operationId": "deleteAllKey",
                "parameters": [
                    {
                        "name": "room_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "acca7421-d920-4751-b227-a21ef604f655"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    }
                }
            }
        },
        "/api/key-encryption/rooms/{room_id}/users/{user_id}": {
            "get": {
                "tags": [
                    "keyEncryption"
                ],
                "description": "get detail key encrytion by room uid and user uid",
                "operationId": "c9533eab9fd31da3e056c51511e41e72",
                "parameters": [
                    {
                        "name": "room_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "acca7421-d920-4751-b227-a21ef604f655"
                        }
                    },
                    {
                        "name": "user_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "acca7421-d920-4751-b227-a21ef604f655"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Trả về  key encrytion của 1 user trong room"
                    }
                }
            },
            "delete": {
                "tags": [
                    "keyEncryption"
                ],
                "description": "Xoá Key",
                "operationId": "deleteKey",
                "parameters": [
                    {
                        "name": "room_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "acca7421-d920-4751-b227-a21ef604f655"
                        }
                    },
                    {
                        "name": "user_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "acca7421-d920-4751-b227-a21ef604f655"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CreateKeyEncryptionSchema": {
                "description": "Create Key Encryption User schema",
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "user_id": {
                                    "nullable": false,
                                    "description": "User uid của nextcloud",
                                    "type": "string",
                                    "example": "acca7421-d920-4751-b227-a21ef604f655"
                                },
                                "room_id": {
                                    "nullable": false,
                                    "description": "Room uid",
                                    "type": "string",
                                    "example": "acca7421-d920-4751-b227-a21ef604f655"
                                },
                                "key_encryption": {
                                    "nullable": false,
                                    "description": "key encryption",
                                    "type": "string",
                                    "example": ""
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
            "KeyEncryptionSchema": {
                "description": "Key Encryption schema",
                "properties": {
                    "user_id": {
                        "nullable": false,
                        "description": "User uid của nextcloud",
                        "type": "string",
                        "example": "acca7421-d920-4751-b227-a21ef604f655"
                    },
                    "room_id": {
                        "nullable": false,
                        "description": "Room uid",
                        "type": "string",
                        "example": "acca7421-d920-4751-b227-a21ef604f655"
                    },
                    "key_encryption": {
                        "nullable": false,
                        "description": "key encryption",
                        "type": "string",
                        "example": ""
                    }
                },
                "type": "object"
            },
            "UpdateKeyEncryptionSchema": {
                "description": "Update Key Encryption User schema",
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "user_id": {
                                    "nullable": false,
                                    "description": "User uid của nextcloud",
                                    "type": "string",
                                    "example": "acca7421-d920-4751-b227-a21ef604f655"
                                },
                                "room_id": {
                                    "nullable": false,
                                    "description": "Room uid",
                                    "type": "string",
                                    "example": "acca7421-d920-4751-b227-a21ef604f655"
                                },
                                "key_encryption": {
                                    "nullable": false,
                                    "description": "key encryption",
                                    "type": "string",
                                    "example": ""
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
...
        }
    },
    "security": [
        []
    ]
}